lint

+++ chia/wallet/cat_wallet/cat_wallet.py
@@ -434,7 +434,7 @@ class CATWallet:
         Returns a set of coins that can be used for generating a new transaction.
         Note: Must be called under wallet state manager lock
         """
-        spendable_amount: uint64 = await self.get_spendable_balance()
+        spendable_amount: uint64 = uint64(await self.get_spendable_balance())
         spendable_coins: List[WalletCoinRecord] = await self.get_cat_spendable_coins()
 
         # Try to use coins from the store, if there isn't enough of "unused"
